:muscle: Rename `inspect.ts` to `_inspect.ts`
